Stripped the header part
